requirements.txt
alembic==1.4.2
click==7.1.2
croniter==0.3.34
dominate==2.5.1
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-RQ2==18.3
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
natsort==7.0.1
pymediainfo==4.2.1
python-dateutil==2.8.1
python-editor==1.0.4
python-magic==0.4.18
python-magic-bin==0.4.14
redis==3.5.3
rq==1.5.0
rq-scheduler==0.10.0
six==1.15.0
SQLAlchemy==1.3.18
visitor==0.1.3
Werkzeug==1.0.1
